requirements.txt
# Core dependencies for autogen exercises and examples
python-dotenv>=1.0.0
duckduckgo-search==8.1.1
agent-framework[azure]==1.0.0b260212
azure-identity>=1.12.0
# Code linting and formatting tools. Not necessary for running the code.
# ruff